chore: interfaces hardening

+++ packages/backend/src/controllers/investments/transactions/get-transactions.controller.ts
@@ -12,8 +12,8 @@ const schema = z.object({
       portfolioId: recordId().optional(),
       securityId: recordId().optional(),
       category: z.nativeEnum(INVESTMENT_TRANSACTION_CATEGORY).optional(),
-      limit: z.coerce.number().min(1).max(100).optional().default(20),
-      offset: z.coerce.number().min(0).optional().default(0),
+      limit: z.coerce.number().int().min(1).max(100).optional().default(20),
+      offset: z.coerce.number().int().min(0).optional().default(0),
     }),
   ),
 });
